Using correct target in rubocop

Closes #408

See merge request noosfero/noosfero!1726

+++ plugins/lattes_curriculum/lib/ext/person.rb
@@ -1,15 +1,14 @@
-require_dependency 'person'
+require_dependency "person"
 
 class Person
-
   attr_accessible :lattes_url, :academic_info_attributes
 
   has_one :academic_info
 
   after_destroy do |person|
     if !person.environment.nil? &&
-person.environment.plugin_enabled?(LattesCurriculumPlugin) &&
-!person.academic_info.nil?
+       person.environment.plugin_enabled?(LattesCurriculumPlugin) &&
+       !person.academic_info.nil?
       person.academic_info.destroy
     end
   end
@@ -22,7 +21,7 @@ person.environment.plugin_enabled?(LattesCurriculumPlugin) &&
     end
   end
 
-  def lattes_url= value
+  def lattes_url=(value)
     if self.environment && self.environment.plugin_enabled?(LattesCurriculumPlugin)
       self.academic_info.lattes_url = value unless self.academic_info.nil?
     end
